% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{cr_get_ratio}
\alias{cr_get_ratio}
\title{Get Contrast Ratio of Two Colors}
\usage{
cr_get_ratio(col_1, col_2, quiet = FALSE, view = FALSE)
}
\arguments{
\item{col_1}{Six-digit hex value preceded by '#', or a named color from
\code{\link[grDevices]{colors}}}

\item{col_2}{Six-digit hex value preceded by '#', or a named color from
\code{\link[grDevices]{colors}}}

\item{quiet}{Whether to print warning when the ratio value is lower than 4.5.}

\item{view}{Whether to plot a demonstration of col_1 text on a col_2
background, and vice versa, for visual inspection. Uses
\code{\link{cr_view_contrast}}.}
}
\value{
A double.
}
\description{
Calculate the color contrast ratio of two provided colors. Intended as an
visual accessibility aid when selecting a text color to place over a single-
color background. The output value should be 4.5 or higher to ensure
sufficient contrast and readability. This function is currently based on
\href{https://www.w3.org/TR/WCAG21/}{WCAG 2.1}. See
\href{w3.org/TR/WCAG/#dfn-relative-luminance}{WCAG} for more information
about the calculation.
}
\examples{
cr_get_ratio("#FFFFFF", "white")
}
