% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{cr_view_contrast}
\alias{cr_view_contrast}
\title{Plot a Demo of User-Supplied Color Pair}
\usage{
cr_view_contrast(col_1, col_2)
}
\arguments{
\item{col_1}{Six-digit hex value preceded by '#', or a named color from
\code{\link[grDevices]{colors}}.}

\item{col_2}{Six-digit hex value preceded by '#', or a named color from
\code{\link[grDevices]{colors}}.}
}
\value{
Character value or vector.
}
\description{
Plots text of one color on a background of another color, and vice versa.
Used to visualise contrasts.
}
\examples{
cr_view_contrast("yellow", "black")
}
