% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{get_ratio}
\alias{get_ratio}
\title{Get contrast ratio of two colours}
\usage{
get_ratio(col1, col2, quiet = FALSE)
}
\arguments{
\item{col1}{Hex value preceded by '#' or a named colour.}

\item{col2}{Hex value preceded by '#' or a named colour.}

\item{quiet}{Whether to print warning when the ratio value is lower than 4.5.}
}
\value{
A double.
}
\description{
Calculate the colour contrast ratio of two provided colours. Intended as an
visual accessibility aid when selecting a text colour to place over a block-
coloured background. The output value should be 4.5 or higher to ensure
sufficient contrast and readability. This function is currently based on
\href{https://www.w3.org/TR/WCAG21/}{WCAG 2.1}. See
\href{w3.org/TR/WCAG/#dfn-relative-luminance}{WCAG} for more information
about the calculation.
}
\examples{
get_ratio("#FFFFFF", "white")
}
