% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{cr_choose_bw}
\alias{cr_choose_bw}
\title{Choose White or Black for Text Overlaying Supplied Color}
\usage{
cr_choose_bw(col_bg)
}
\arguments{
\item{col_bg}{Hex value preceded by '#' or a named color.}
}
\value{
A character value: "white" or "black".
}
\description{
Given a background color, what's a better color for overlay text: white or
black? Calculated as per \code{\link{cr_get_ratio}}. Defaults to black in the
case of a tie.
}
\examples{
cr_choose_bw("grey90")
}
